requirements.txt
# Production requirements for GigaMind - Compatible versions
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.6
fastapi==0.104.1
pydantic==2.4.2  # Compatible with fastapi 0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
gradio==4.7.1
huggingface-hub==0.17.3
certifi
urllib3>=2.0.0
sqlalchemy==2.0.23
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.2
accelerate==0.24.1
bitsandbytes==0.41.3
torch==2.1.1
vaderSentiment==3.3.2
numpy==1.24.3
scikit-learn==1.3.0
# Additional compatible dependencies
typing-extensions>=4.8.0
anyio>=3.7.1
